models directory is created
